Added a table of contents, finished commits, added branch management
Added a table of contents that links to each chapter and section to what is shown. This can be linked to any section header.

Finished up commits, talking about when to make a commit and when to push, as code often does not get pushed when clocking out at the end of the day. The added sections outlines what should be done in those cases.

Added a section for managing branches. This covers how to properly manage a branch, when to merge, and what processes are involved when merging a branch. Pull requests and directly merging into the main branch are listed as methods.

Other minor changes include:

- Linting document to follow markdown rules
- Changing headers to make sense in the document
- Removed "work in progress" tag at the top of the document